Support other Override locations in update trust script
Currently, the script to update trust info hardcodes the RecipeOverrides dir to ~//Library/AutoPkg/RecipeOverrides. In setups where the AutoPkg preferences specify a different location for RecipeOverrides (via RECIPE_OVERRIDE_DIRS), it should be honoured. This PR assumes that RECIPE_OVERRIDE_DIRS is a string (not an array, which is supported but highly uncommon). This does not account for the case where the prefs supplied by the user have specified another RecipeOverrides directory (but then again, neither does the current script as far as I can tell).